requirements.txt
torch==2.0.1
torchvision==0.15.2
Pillow==10.0.0
matplotlib==3.7.2
numpy==1.24.3
pyyaml==6.0.1
scikit-learn==1.3.0
seaborn==0.12.2
gradio==3.50.2
